# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Tábla beadando Reszt full API ",
    "description": "Mérkőzéses webes APIs."
  },
  "host": "127.0.0.1:3000",
  "basePath": "/",
  "schemes": ["http"],
  "paths": {
    "/login": {
      "post": {
        "summary": "Bejelentkezés és token generálása",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": { "type": "string" },
                "password": { "type": "string" }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sikeres bejelentkezés! Token mentve.",
            "schema": {
              "type": "object",
              "properties": {
                "token": { "type": "string" }
              }
            }
          },
          "401": { "description": "Te kis huncut! Hibás felhasználónév vagy jelszó..." }
        }
      }
    },
    "/meccsek": {
      "get": {
        "summary": "Minden mérkőzés listázása",
        "responses": {
          "200": {
            "description": "Sikeres válasz",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": { "type": "integer" },
                  "datum": { "type": "string", "format": "date" },
                  "kezdes": { "type": "string" },
                  "tipus": { "type": "string" },
                  "belepo": { "type": "integer" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Új mérkőzés hozzáadása",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "datum": { "type": "string", "format": "date" },
                "kezdes": { "type": "string" },
                "belepo": { "type": "integer" },
                "tipus": { "type": "string" }
              },
              "required": ["datum", "kezdes", "belepo", "tipus"]
            }
          }
        ],
        "responses": {
          "201": { "description": "Mérkőzés sikeresen hozzáadva" },
          "403": { "description": "Nananananaaaaa... nincs is jogosultságod! Lépj be!" }
        },
        "security": [
          { "bearerAuth": [] }
        ]
      }
    },
    "/bajnoki": {
      "get": {
        "summary": "Bajnoki mérkőzések időrendben",
        "responses": {
          "200": {
            "description": "Sikeres válasz",
            "schema": {
              "type": "array",
              "items": { "type": "string", "format": "date" }
            }
          }
        }
      }
    },
    "/belepesek/{meccsid}": {
      "get": {
        "summary": "Nézők lekérdezése adott mérkőzésre",
        "parameters": [
          {
            "name": "meccsid",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Sikeres válasz",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "nev": { "type": "string" },
                  "ferfi": { "type": "boolean" },
                  "berletes": { "type": "boolean" },
                  "idopont": { "type": "string", "format": "date-time" }
                }
              }
            }
          },
          "403": { "description": "Nananananaaaaa... nincs is jogosultságod! Lépj be!" }
        },
        "security": [
          { "bearerAuth": [] }
        ]
      }
    },
    "/utolsojegyar": {
      "get": {
        "summary": "Legutolsó mérkőzés jegyárának lekérdezése",
        "responses": {
          "200": {
            "description": "Sikeres válasz",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "belepo": { "type": "integer" }
                }
              }
            }
          },
          "403": { "description": "Nananananaaaaa... nincs is jogosultságod! Lépj be!" }
        },
        "security": [
          { "bearerAuth": [] }
        ]
      }
    },
    "/meccsek/{id}": {
      "delete": {
        "summary": "Mérkőzés törlése",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": { "description": "Sikeres törlés" },
          "403": { "description": "Nananananaaaaa... nincs is jogosultságod! Lépj be!!" }
        },
        "security": [
          { "bearerAuth": [] }
        ]
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}